+++ cpp/inc/bond/protocol/simple_binary.h
@@ -208,7 +208,7 @@ protected:
 
 
 template <typename Buffer>
-const uint16_t SimpleBinaryReader<Buffer>::magic = ProtocolType::SIMPLE_PROTOCOL;
+const uint16_t SimpleBinaryReader<Buffer>::magic = static_cast<uint16_t>(ProtocolType::SIMPLE_PROTOCOL);
 
 
 /// @brief Writer for Simple Binary protocol
